Removes old code
