--- node/clients/sitemap/base.ts
@@ -1,7 +1,0 @@
-export interface SiteMap {
-  fromLegacy: (forwardedPath: string) => Promise<string>
-
-  replaceHost: (str: string, forwardedHost: string, rootPath?: string) => string
-
-  appendSitemapItems: (currSitemap: any, items: string[]) => void
-}
