--- node/clients/sitemap/index.ts
@@ -1,9 +1,0 @@
-import { SiteMap } from './base'
-
-export const sitemapClientFromCtx = ({
-  vtex: { platform },
-  clients,
-}: Context): SiteMap =>
-  platform === 'gocommerce'
-    ? clients.sitemapGC
-    : clients.sitemapPortal
