--- node/middlewares/state.ts
@@ -1,5 +1,0 @@
-export async function prepareState(ctx: Context, next: () => Promise<any>) {
-  ctx.state.platform = ctx.vtex.platform || 'vtex'
-
-  await next()
-}
