+++ node/service.json
@@ -6,48 +6,8 @@
   "maxReplicas": 10,
   "workers": 4,
   "routes": {
-    "sitemapXML": {
-      "path": "/sitemap.xml",
-      "public": true
-    },
-    "departments": {
-      "path": "/sitemap-departments.xml",
-      "public": true
-    },
-    "brands": {
-      "path": "/sitemap-brands.xml",
-      "public": true
-    },
-    "custom": {
-      "path": "/sitemap/sitemap-custom.xml",
-      "public": true
-    },
-    "user": {
-      "path": "/sitemap/sitemap-user-routes.xml",
-      "public": true
-    },
-    "products": {
-      "path": "/sitemap-products-:page:.xml",
-      "public": true
-    },
-    "categories": {
-      "path": "/sitemap-categories.xml",
-      "public": true
-    },
-    "category": {
-      "path": "/sitemap-category-:specifier:.xml",
-      "public": true
-    },
-    "sitemap": {
-      "path": "/sitemap/:path",
-      "public": true
-    },
-    "newSitemap": {
-      "path": "/_v/public/newsitemap/:path",
-      "public": true
-    },
     "newSitemapWithBinding": {
-      "path": "/_v/public/newsitemap/:binidingIdentifier/:path",
+      "path": "/:binidingIdentifier/sitemap/:path",
       "public": true
     },
     "generateSitemap": {
@@ -57,11 +17,6 @@
     "robots": {
       "path": "/robots.txt",
       "public": true
-    },
-    "canonical": {
-      "path": "/canonical",
-      "public": false,
-      "smartcache": true
     }
   },
   "events": {
